pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "submission-analyzer"
version = "0.1.0"
description = "CLI tool to analyze Sherlock and Code4rena contest submissions"
requires-python = ">=3.10"
dependencies = [
  "requests",
  "python-dotenv",
  "python-telegram-bot",
  "sentry-sdk"
]

[project.scripts]
sherlock-analyzer = "submission_analyzer.platforms.sherlock.main:main_sync"
code4rena-analyzer = "submission_analyzer.platforms.code4rena.main:main_sync"
